requirements.txt
appdirs==1.4.3
attrs==19.3.0
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
black==19.10b0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.9.2
Django==2.2
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-restframework==0.0.1
django-storages==1.9.1
djangorestframework==3.11.0
idna==2.9
pathspec==0.7.0
pkg-resources==0.0.0
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
regex==2020.2.20
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
